Remove Non-Batman Issue from Reading List
This list was originally based on one of TPBs, and the TPB for Batman: Irresistible contained the Hourman comic at the end.  The reason for its inclusion, however, was because of the common artist.  It doesn't include Batman, has no relevance to his world, and isn't even that useful as a standalone story as it's the beginning of a multipart.  Ironically, the NEXT issue in the Hourman series features a Batman (of many centuries hence).  I don't think this issue belongs here.