Space Goat's brief, but productive Evil Dead 2 run
Space Goat for a short time had the license to the Evil Dead 2 comic rights, and boy did they shoot some books out in that 2 year span! Tales of Ex-Mortis is mostly self contained one shot issues not really in the continuity but as it's the only release they had not explicitly in the same continuity I threw it in for a complete reading of the company's Evil Dead offering.