Friday the 13th Master List
There's no overarching story, but slasher fans are used to this and often we want everything seriously a part of any of the continuities of our franchises. So with that in mind, here's every Friday the 13th-specific book, only omissions for Jason's appearances in comics are the comedic ones in parody books. 

This is all in release order as again, no story to really follow but there's no real comprehensive list like this out there so I pieced it together from CV and my knowledge of this franchise (my first two movies ever watched were Happy Hooker Goes to Hollywood and Friday The 13th Part 3, one could say I destined to be this way) as a literal lifelong fan.