Added TMNT Laird Canon Volume 1
TMNT has a confusing and fractured history, even prior to IDW. Peter Laird has gone on record to say he only considers the stories by him and Eastman to be canon from the original volume 1 and Tales volume 1. I've compiled this into a chronological order story-wise with a deviation of #24-26 they are referenced in the Laird/Eastman books after so it must be canon even by his standard.