Grimm Fairy Tales Other Stories Act 1
These are non-main stories but still firmly part of the GFT series, stuff like april fools or series that spun off from the main one but didn't sprawl as much as Wonderland did. These will be concurrent with the Main and Wonderland lists in chronology but this set will be fairly optional.