Zenescope folder plus 2 acts
Since there's going to be a large amount, created a Zenescope folder with 3 sub-folders for each set of Reading Lists my source is providing. The main series, Wonderland, and the non-canon/smaller spin-offs.

Source has 9 acts that bring us to about 2019 so after that we'll be relying on my reading through the 2019+ stories for a guide.