Update Adam Warlock.cbl
Warlock #8 came out in Oct 1973. Warlock #9 came out in Oct 1975. The storyline from #8 was finished in The Incredible Hulk 176-178. Next Jim Starlin began writing Warlock stories in Strange Tales 178-181. Strange Tales 181 came out in Aug of 1975. After that Warlock #9 was advertised as a Premiere Issue and came out in Oct 1975.

This pull request is to move those Hulk and Strange Tales issues into the correct chronological AND story order between Warlock #8 and #9.