Add files via upload
A More streamlined reading order for secret wars, omitting a lot of unnecessary/optional tie-ins that may scare off newer readers (like myself) looking to read the event in it's entirety. This was mainly curated with help by CBH, but perfecting the order as well as the chronology was somewhat of a personal endeavor. 

https://www.comicbookherald.com/the-complete-marvel-reading-order-guide/secret-wars-reading-order/
https://www.comicbookherald.com/the-complete-marvel-reading-order-guide/secret-wars-reading-order/battleworld-warzone-last-days-checklists/